allow runtime changes to batchDelay
